[Android] Make the uid that can connect to the devtools server configurable
Currently, the devtools server allows only code running in the same process
to connect. This change allows specifying the process that can connect to
devtools server.

This code is being modified since the XDK architecture for remote debugging
has changed. Previously, App Analyzer had a websocket proxy that used to run
in the same process as xwalk. In the next version of XDK, the websocket proxy
is being moved to be part of another application (App Preview) and will run
in a process that is different from xwalk. With the addition of this API,
App Preview service will instruct App Analyzer to allow connections from
the service.

BUG=https://crosswalk-project.org/jira/browse/XWALK-964