[Android] Fix packaging tool fails to build application when manifest.json doesn't include "icons" object.
In the manifest_json_parser.py, "icons" object with multiple entries in manifest.json
is represented as a dictionary, so it should not be a string even if it's empty.

Packaging tool would fail when iterating the dictionary if "icons" object was unset
in manifest.json.

BUG=https://crosswalk-project.org/jira/browse/XWALK-909

This patch is cherry-picked from 939c7e50fb82.