Remove AUTHORS.
It has never been properly enforced and it does not have any effect for us.

We don't require a CLA, and we don't mention "The Crosswalk Authors" or
anything like that in the copyright headers on files.

Anyone can just send their patches and add their names as copyright holders
to the headers in the files they touch if the contributions are big enough.