[Android] Exit with error code when the make apk test cases failed.
Because the trybot requests exit code for main page to check
whether PR passes the make apk test cases. Currently,
the make_apk_test.py only display the error info, but always exit
with 0 even if it failed. So the main page of trybot will always show
tests passed.

Add a function to check the make_apk_test.py's result, and exit
with correct code according to this result.