Sync gyp_xwalk with the current contents of gyp_chromium.
gyp_chromium underwent a large amount of changes since M32 due to the work
on GN upstream. This broke the Android build (gyp_xwalk did not even detect
it was supposed to set up an Android build in the first place).

I have basically copied gyp_chromium on top of gyp_xwalk and applied the
relevant changes after comparing gyp_xwalk with M32's gyp_chromium, with the
addition of a change that does not run GN when using the make generator (as
a temporary measure to keep the Tizen build working while we find a better
solution).