[Android] Fix some Java compiling warnings for Crosswalk.
Since some webkit APIs and Android OS APIs are deprecated as the
updated Android OS version, there are compiling warnings in Crosswalk.
This fix resolves these issues by using specific modes. For deprecated
parameter type in API, we suppress the warnings, for example, in
XWalkWebChromeClient::onExceededDatabaseQuota(), the parameter type of
WebStorage.QuotaUpdater is deprecated in Android 4.4, while it is still
used before Android 4.4, no other parameter is to replace it. For the
deprecated APIs, the compatible APIs are used to replace them accordingly.

BUG=https://crosswalk-project.org/jira/browse/XWALK-1064