[IVI][Runtime] Fix unable to install application on Tizen IVI 3.0 image.
The package install/uninstall/update for Tizen are designed to integrate
with AMD (Application Management Daemon) on Tizen mobile 2.x, not
compatible with Tizen IVI 3.0.

This is a temporary solution to revert to use macro OS_TIZEN_MOBILE, this
means install/uninstall/update won't work with AMD on Tizen IVI. We'll
add that support back later.

BUG=XWALK-1074