[Android] Fix the regression issue of http authentication on chromium 34.
Based on chromium 34, the http authentication cannot work, while
it works on chromium 32, this is because Chromium 34 starts to
support the out-of-process iframe, RenderViewHost is replaced by
RenderFrameHost, while in XwalkLoginDelegate, webcontents is still
gotten via the RenderViewHost, which causes the corresponding UI
thread for http authentication not to work. This fix resolves this
issue by replacing RenderView with RenderFrame for http authentication.

BUG=https://crosswalk-project.org/jira/browse/XWALK-1088