[Android] Package xwalk.apk as resource in xwalk_core_library
This is step 2 of getting rid of assets in xwalk_core_library.
To use paks as res/raw, the idea is to generate an xml in res/values
when creating xwalk_core_library, the xml contains a string-array
to describe all the paks will be put in res/raw.
XWalkViewDelegate will try to load the xml first, and intercept
ResourceExtractor if the resource to extract is on the list.

BUG=https://crosswalk-project.org/jira/browse/XWALK-1085