Flaky crashes of XWalkExtensionsIFrameTest tests
The problem here was due to a race condition with
XWalkExtensionServer. The XWalkExtensionServer methods
are invoked via task runner (on message loop), on the
other hand XWalkExtensionServer is deleted on message
loop, so the crashes turned up when the deferred calls
were processed after the object deletion.

Fixed now using WeakPtrs.