[Tools] Add scm-remove-wrong-checkout.py.
This is a helper script intended to be invoked via "gclient recurse".
fetch_deps.py calls it automatically.

It is intended to remove a directory pointed to by a gclient entry when it
is supposed to be a git checkout but contains an svn one (and vice-versa).

The main idea is to prepare the ground for using our v8-crosswalk
repository: since V8 so far has been cloned from Google's SVN repository, if
we then switch to our git repository gclient sync will fail.