DEPS.xwalk: Bring in v8-crosswalk.
Start tracking v8-crosswalk in src/v8. v8-crosswalk is a V8 fork along the
lines of chromium-crosswalk and blink-crosswalk. For now its main purpose is
to have all the SIMD.js optimizations some folks have been working on. Other
than that, it should be the usual V8 we all know.