[Tizen] Get rid of the crosswalk-emulator-support package.
crosswalk-emulator-support was used during the Tizen 2.x times because the
Tizen emulator's GL implementation did not work as expected and we thus had
to fall back to libosmesa.

Now that Tizen 2.x is a thing of the past, we can get rid of it. If the
Tizen 3 emulator still has problems, we are now in the position to try to
fix it.