[Android] Refine the callback function of compressor.
This patch is to refine the feature of compressor.
optparse stores the converted value when we use '--compressor'
at the first time. Then we use '--compressor=css', optparse
does not consume the argument.
Add 'type' and 'nargs', they instruct optparse to consume the
argument and convert the parameter to related type.
Add test case for this feature.

BUG=https://crosswalk-project.org/jira/browse/XWALK-1151