[Android] Provide embedding API for Crosswalk.
A new embedding API is designed for providing capabilities to embedders.
This can provide functionalites and their APIs as an Android library.
This patch mainly re-define API and re-construct classes and their
relationship. To keep current capability, lots of legacy code are also kept
without big changes. Main changes are:
1. New APIs: XWalkView, XWalkUIClient, XWalkResourceClient, XWalkPreferences,
XWalkNavigationHistory, XWalkNavigationItem and XWalkJavascriptResult.
2. Legacy APIs are used by tests. Need to refine them.
3. Remove unused code for http authentication.
4. Re-construct some duplicated code in core tests.
This issue is still for XWALK-1041. It's not finished yet. Need to implement
remaining TODOs and refine the description of APIs.