[Android] Enable the immersive fullscreen with manifest.
The fullscreen setting in manifest doesn't work with Android 4.4,
this is because there is no way to declare immersive mode in the
manifest or window flags. This fix resolves this issue by checking
the manifest fullscreen setting via manifest parser and calling the
corresponding functions of setting the immersive fullscreen mode.
In addition, the command-line parameter of immersive fullscreen
mode is also supported for packaging tool.

BUG=https://crosswalk-project.org/jira/browse/XWALK-1200