[Android] Update the package name handler in packaging tool.
In packaging tool, if the option of package name is not set or
the the wrong package name is set, there is no information prompt.
in addition, when there is space in activity name or package path,
the produced package cannot be launched, this is because the Dex
used in packaging tool cannot work with package path including space,
there is a launching related file is lost. And also javac.py cannot
work with the specific illegal characters. This fix resolves these
issues by updating the handler of package name.

BUG=XWALK-1210