[SysApps][Android] Support lastUpdated in Contacts API
1. When new a contact, current time will be saved into the contact.
2. When update a contact, we let underlying database to handle the last
updated timestamp. This is the only way to keep it right as we can't
predict other apps of modifying system's contacts.
3. When function of find is called, the last updated timestamp field
will be fetched from the database and inserted into the results.
4. Only Android version >API 18 will be supported, as the last updated
timestamp field in the database was first introduced in API 18.
5. For lower Android versions, the lastUpdated timestamp will be seen
when creating a new contact, but it will not be saved or returned in
search results.

BUG=XWALK-1028