[Android] Add a new flag for universal file access
Add a new general API for settings, with the format:
    void setSettingEnabled(String name, boolean value);

Add a new command line of "--allow-universal-access-from-file-urls"
in make_apk.py, by which JavaScript running in the context of a
file scheme URL could access content from any origin.

Change back to use file scheme temporarily to package apps due to
app scheme is not recognized as a local one in blink.

BUG=XWALK-1256