[Tizen] NativeAppWindowTizen should be used in Tizen IVI
NativeAppWindowTizen should be used in Tizen IVI as it contains some
logic that is needed there, e.g. implementation of MultiOrientationScreen
interface (which is used by screen orientation extension).

Before the change an attempt to use screen orientation API caused an
xwalk daemon crash, this patch fixes the following bugs:

BUG=XWALK-1192
BUG=XWALK-1141