[Tizen] Greatly simplify the handling of incremental builds.
Also known as "it takes weeks to make things simpler".

The behavior for users should be the same (eg. gbs build --define
BUILDDIR_NAME), but the way we handle it in the spec file is very different.
Instead of doing a lot of black magic with gyp flags that work around some
very specific behavior of gyp's make generator, just create a symlink to
BUILDDIR_NAME when it is specified.

In addition, simplifying the code and making it less dependent on the make
generator also allows us to support incremental builds with ninja, which is
going to be the only supported build system once we move to Chromium 35.