[Extensions] Add command line switch xwalk-extension-cmd-prefix
This will prepend the contents of this flag to the command line
when starting extension process, it works as same as the switch
"--renderer-cmd-prefix", could be useful if we need to debug
the extension process at its very beginning.
Useful values might be "valgrind" or "xterm -e gdb --args".

Also modified the child path flag which is used to retrieve the
child process' execution file according to the extension command
line prefix on linux. In most cases we need to use CHILD_ALLOW_SELF
to exec child process from /proc/self/exe to prevent chromium's
autoupdate feature to change the file out from underneath; For the
case of xwalk-extension-cmd-prefix, as we suppose to override the
command line to load the app directly, we need to use CHILD_NORMAL
to exec child process from its real app path.