gyp_xwalk: Stop running GN.
GN is not a hard dependency for generating the build files in M34, and add
unnecessary checking for dependencies in the host system when we're building
for Android and Tizen.

In addition, the hack in place there to avoid running it with the make
generator will be rendered useless once we start using Ninja on Tizen.