[Tizen] Start using Ninja to build Crosswalk on Tizen.
All Tizen profiles finally have Ninja as a package, and commits 096e68d and
9c9419b did all the required heavy-lifting to allow us to use the ninja gyp
generator instead of the make one to build the Crosswalk RPM.

Incremental builds should still be supported just fine, but the first build
with ninja will start from scratch due to the move to a different build
system.

One meaningful change is that we do not pass -Duse_system_icu=1 anymore and
instead use Chromium's copy. While that is not perfect from a packaging
perspective, it is currently necessary due to problems with the code (see
crrev.com/104793004 and crrev.com/124003004).

Moving to Ninja is a required step for us to move to Chromium 35, since the
make gyp generator is unsupported from this version on.

BUG=XWALK-1025