M35 Gardening.
Mac was not running and had a fatal check at launch. This commit
fixes the problem by copying the newly icudat*.dat to the Mac bundle.
Also make sure that the override of the bundle path is done at the
right moment and this before the PreSandboxStartup for it to work.

Upstream CL : https://codereview.chromium.org/109013004