[Tizen] packaging: explicitly disable GNOME Keyring support.
gnome-keyring isn't packaged for Tizen, and we are not interested in using
it at the moment anyway.

While support for it is disabled by default when Aura is used in M34, this
is not the case anymore since Chromium r251992 (part of M35). Prepare for it
by explicitly disabling the support in master so that the actual commit that
moves to M35 in crosswalk is smaller.