[Android] Try to match cpu arch after loading xwalk core
It's extremely slow to load ARM binaries on ia devices.
And it hurts user experience. Add a method in libxwalkcore.so
to tell XWalkViewDelegate whether the binary is built for IA.

The exception of CPU architecture mismatch will be handled
by XWalkView instead of dispatch to runtime level or embedders.

A AlertDialog will be shown to notify user what's wrong and
suggest them report feedback to application developers or
rate and comment the application on PlayStore.

BUG=https://crosswalk-project.org/jira/browse/XWALK-1283