[Tizen IVI] Support landscape window orientation by default
Before the change 'NativeAppWindowTizen' hardcoded portrait
orientation to be the default one, however (unlike the mobile
devices) the default window orientation on IVI devices is
landscape. Thus the wrong window transformation was applied
causing bugs, e.g. XWALK-1313.

This patch adds support (not complete yet, but enough for most
of the applications) for the default landscape window orientation.

BUG=XWALK-1313