[Tizen][Runtime] Add support for 'hwkey' in Tizen widget configuration.
Tizen WRT Core Spec shows that when hardware key(‘back’, ‘menu’) is
pressed, WRT MUST send the “tizenhwkey” custom event to the Web
Application if the <tizen::setting hwkey=”disable”/> element is not
specified. Therefore, app developer can add customized behavior for
this.

Spec:
https://source.tizen.org/sites/default/files/page/tizen-2.2-wrt-core-spec.pdf
(item 0470)

This implementation will fetch event from OzoneMessagePump, if it is a
key event, and at the same time the 'hwkey' flag is not disabled in
widget configuration, browser process will send the event to render process,
so that RP can handle the hardware key event.

BUG=XWALK-1166