[Runtime] Support to load application resources by the locale setting.
For WGT package, the W3C widget specification defines the concept of
folder-based localization, which is the process whereby an author places
files inside folders that are named in a manner that conforms to a
language-range ABNF rule. That is, by naming folders in lower-case using
values derived from the IANA Language Subtag Registry such as "en-us",
"en-gb", and so on. These locale folders are then placed inside the
container for localized content which is a reserved folder at the root
of the widget package whose folder-name case-sensitively matches the
string 'locales'.

For more details, please refer to
http://www.w3.org/TR/widgets/#folder-based-localization.

The related feature is XWALK-1030 and XWALK-1503