[Runtime] Support identifying namespace of content element in WGT config.xml.
In Tizen platform, there are two kinds of element to set the application
start page in configure document (config.xml), the <content> and
<tizen:content>, the content element with tizen namespace represents the
host app starting page, the other one is for common packaged app.

This patch can choose the right one for packaged app, the tizen specific
one will be ignored when launching from local page.

This patch will support to get the first one if multiple content elements
are defined in config.xml as well.