[Android] Hide classes and description refinement for Javadoc.
This fix aims to use Doclava(https://code.google.com/p/doclava/) to
generate java doc for embedding API. Since some classes and methods
are public but not generated as API, "@hide" is used from Doclava.
Besides, it refines some description and comments. Next is to enable
doclava in Crosswalk.

BUG=https://crosswalk-project.org/jira/browse/XWALK-1544