[Android] Allow XWalkUIClient to overwrite openFileChooser
Previous, select file is handled within xwalk_web_contents_delegate
and the openFileChooser API in XWalkUIClient is not working.
"openFileChooser" is used in android.webkit.WebChromeClient, which
is adopted by many current webview users (e.g. Cordova).

In this commit, it dispatches the request to XWalkUIClient which is
empty implementation. If the derives override the implementation,
it can invoke the callback somewhere to provide the chosen file.
Otherwise, it will fallback to internal solution using
RuntimeFileSelectHelper::RunFileChooser.

BUG=https://crosswalk-project.org/jira/browse/XWALK-1561