[Android] Enable reflection layer generator
Add a python script to generating code for classes in
org.xwalk.core from classes in org.xwalk.core.internal
which need to be exposed in Embedding API.

For each class of org.xwalk.core.internal which needs to be
exposed in Embedding API.
A bridge class and a wrapper class will be created for it.
The bridge class is a subclass of internal class and maintains
a wrapper object of this type.
A wrapper class is in org.xwalk.core and be exposed in Embedding
API, it maintains an instance of bridge object as well.
Wrapper and bridge talks to each other via pure reflection. So
they don't have to be in the same package in future.

TODO: JavascriptInterface class needs to be placed in wrapper and
pass the class object to internal for recognition of javascript
interface.

BUG Related: https://crosswalk-project.org/jira/browse/XWALK-1638