Bump Crosswalk to 36.0.1985.18
This commit bump crosswalk to the newer M36 milestone.

Important commits :
- Sync gyp_xwalk with upstream gyp_chromium r263984.

- Update envsetup.sh due to upstream changes.
build/android/envsetup.sh from upstream renamed some variables. Make
sure we use these ones now.

- WebAudio is now enabled on x86 Android by default.

- [Tizen][Runtime] Fix MessagePumpOzone issue for M36 rebase.

MessagePumpOzone and MessagePumpObserver are deprecated in M36.
This patch removes related codes in Crosswalk in order to make it
compilable for M36 rebase.

Effect of this change is temporarily 'hwkey' feature doesn't work.

The following fix patch will be provided once we have successfully
switched to M36.

See bug XWALK-1420

-[Android] M36 Gardening : Refactor the notification system

The notification code needs to be refactored due to upstream changes.
Few methods names were renamed or removed and the biggest change was
the removal of the notification id that we were using.

This patch essentially adapt to the changes (notably passing the delegate
instead of process_id and route_id) and it rolls our own way to manage
notifications ids.

Upstream CL : https://codereview.chromium.org/241553002

- DEPS.xwalk: Roll ozone-wayland to a5ca2e9.

- [Tizen] Add crosswalk-do-not-look-for-gtk-dependencies-on-x11.patch.

Removing crosswalk-do-not-look-for-gtk2-when-using-aura.patch was not
enough, as use_ozone==0 in the Tizen Mobile (ie. X11) case and gyp was
then looking for GTK+ dependencies.
Add a patch that also checks for tizen==0, but only for Mobile, since
the other profiles do not need it.