[Android] Refine the argument processing for compressor.
This patch is to refine the argument processing for compressor.
In prior version, the packaging process will not be terminated when the
option value for compressor is wrong.
Now use the "choices" parameter, if the argument is not in the acceptable
value list, provide the advice to user to refine the option.

BUG=XWALK-1672