Fix processing of the Manifest 'start_url' member
Fix processing of the Manifest 'start_url' member
after resolving an ambiguousness in the spec
https://github.com/w3c/manifest/pull/200/files

From now 'start_url' manifest key is an application
entry point.

So far 'start_url' cannot contain only a relative URL
(the resulting URL is constructed using the path to
the manifest file as a base URL:
http://w3c.github.io/manifest/#dfn-steps-for-processing-the-start_url-member)

External URLs handling should be clarified in the spec
further.