[Next][Android] M36 Gardening : Refactor the notification system
The notification code needs to be refactored due to upstream changes.
Few methods names were renamed or removed and the biggest change was
the removal of the notification id that we were using.

This patch essentially adapt to the changes (notably passing the delegate
instead of process_id and route_id) and it rolls our own way to manage
notifications ids.

Upstream CL : https://codereview.chromium.org/241553002