[Tizen][Runtime] Fix can not update application.
In the old logic, when we update an application, we
uninstall the old application and install new application,
but when we call
     <pkgmgr_parser_parse_manifest_for_installation>
and  <pkgmgr_parser_parse_manifest_for_uninstallation>
on one process, this process will crash.So we should use
     <pkgmgr_parser_parse_manifest_for_upgrade>
to update.

And we should make a new pkg start point for new app data.

BUG=XWALK-1738