[Android] Add test case for onLoadStarted() and onLoadFinished().
This patch is to add test case about onLoadStarted() and onLoadFinished().
For onLoadStarted: Set url to iframe, load data from webserver, get the
url when onLoadStarted() was called, compare it with the expected value.
For onLoadFinished: Set up a url which contain iframe, load this url,
get the url when onLoadFinished() was called, compare it with the
expected value.