[Android] Add test case for setValue() and getValue().
This patch is to add test case for setVaule() and getValue() in
XWalkPreferences.java.
Set the value with key and enabled value, get the value by the key,
compare it with the excepted result.
If the key is not supported, it will throw runtime exception, catch
it and make sure it's the instance of RuntimeException.