[Android] Fix the repeat copy issue when app name equals to app root.
This patch is to fix the issue about repeat copy when the app name
equals to app root.
In the function "Prepare()" of customize.py, if the app name equals
to app root, it will remove the directory according to the app name,
so app root was not existed. It will lead to repeat copy.
If the app name equals to app root, add prefix to app name to resolve
this issue.

BUG=XWALK-1703