[Runtime][Tizen] Fix hwkey issue after M36 rebase.
Crosswalk fetches hardware key events from MessagePumpOzone before.
However, MessagePumpObserver is deprecated in M36, it is replaced by
PlatformEventObserver, which is installed on a PlatformEventSource.
Besides, the ozone message pump is also removed in M36 and is
substituted by the default message pump based on libevent instead.

This patch fixes above issue and now we support 'hwkey' feature again.

BUG = XWALK-1420