[Android] Add samples for hasEnteredFullscreen() and leaveFullscreen().
This patch is to add samples for hasEnteredFullscreen() and
leaveFullscreen().
For video: create a html file, it contains video tag. Open this html file,
click the button for full screen, the video will be displayed in
fullscreen. Press the return button, video will be exited the fullscreen
mode.
For html5 api: click a button to execute the javascript for fullscreen,
then click button to exit the fullscreen mode.