[Android] Pre-initialize the CPU information for ARM platform
On ARM platform, the cpu information is fetched from /proc/cpuinfo
which requires the right to allow accessing IO. Howerver, in Crosswalk,
the IO/IPC thread is not allowed to IO access unless
ThreadRestriction::SetIOAllowed(true) is called explicitly.

This fix is to pre-initialize the CPU info in advance to avoid thread
restrictions on reading /proc/cpuinfo.

BUG=https://crosswalk-project.org/jira/browse/XWALK-1836