Use .DEPS.git instead of DEPS
This commit is allow developers get GIT version of dependencies.

Most of dependencies in <release_dir>/DEPS are pointing to subversion.
Unfortunately, svn command behind firewall randomly fails with error
"501 Not Implemented". Instead, git repos are more stable to fetch.

Since upstream is switching SVN to GIT, it is time to switch to
GIT(.DEPS.git) totally.

Related to XWALK-1894