[Android] Add test case for onActivityResult() and onNewIntent().
This patch is to add test case for onActivityResult() and onNewIntent().
For onActivityResult: In the main activity, call startActivityForResult()
with request code, then in auxiliary activity, set result with result code
and extra data. In main activity, verify the request code, result code and
extra data.
For onNewIntent: In AndroidManifest.xml, set the launch mode to "singleTask"
for main activity, then start activity to auxiliary activity, click button
to return to main activity, in onNewIntent(), verify the value and set text
result.