[Android] Fixed android build issue
As upstream bumps to version 36.0.1985.18, we found we always
failed to build xwalk android as we haven't pass '-DOS=android'
to gyp_xwalk. So added it into build/android/envsetup.sh.

Also removed '--check' because if we pass '-DOS=android' to
gyp_xwalk, the '--check' will be appended into args.