[Android] Retain the application name with space on device.
This patch is to retain the application name with space.
Create a class, it contains the attributes of application, set the
value according to the content of manifest or command line.
It also contains the original name and replaced name.
Rewrite the label value in AndroidManifest.xml with original name.
(cherry picked from commit e67562f2547babe07a668f6068528c2e30b6eee0)

Conflicts:

	app/tools/android/customize.py
	app/tools/android/make_apk_test.py