[Android] Retain the application name with space on device.
This patch is to retain the application name with space.
Create a class, it contains the attributes of application, set the
value according to the content of manifest or command line.
It also contains the original name and replaced name.
Rewrite the label value in AndroidManifest.xml with original name.