[Android] Add test case for onHide() and onShow().
This patch is to add test case for onHide() and onShow().
Create a html file, it contains video tag. The video will auto
play when the page was loaded, call onHide(), the video will stop,
call onShow(), the video will play. Set the flag according to the
status of video. Verify the flag with expected result.