[Android] Decouple url loading and onReadyToRender callback
The url loading should not depend on the readiness of compositing
surface. Otherwise, it will cause 'deadlock' between loading url and
surface availability in some cases.

For example, to support splash screen in cordova, the XWalkView should
be initially invisible and waits for a signal from web page to become
visible. However, Android framework may not create the underlying Surface
for an invisibile View, which causes the url won't be loaded.

BUG=https://crosswalk-project.org/jira/browse/XWALK-1884