[Android] Transition from route_id to render_frame_id and child_id to render_process_id.
This patch is to transform the route_id to render_frame_id and child_id
to render_process_id.
Set cache mode does not work without this patch.
The request was handled by the interfaces in "InterceptNavigationDelegate",
since the io client was got by the wrong id, so the interfaces in
"IoThreadClientThrottle" were not called.

BUG=XWALK-1908