[Android] make_apk.py error when Android SDK environment path contain whitespace
The javac.py option --classpath has 2 files, xwalk_app_runtime_java.jar and android.jar,
which paths joined by whitespace. But if one of paths contain whitespace. The javac.py
use options.classpath.split() will parse a error path. So change the 2 paths joined with '|'.
For contain '|' in file name is not recommended in Linux and invalid in Windows.

BUG=XWALK-1905