[Android] Deprecate xwalk_android_gyp
In M36, android_gyp is still available in upstream, but it is
going to be removed in recent future.
We need to deprecate xwalk_android_gyp together with upstream's
change, and we need to pass '-DOS=android' explicity to it to
generate android building environment.

This change based on the discussion in pull request:
https://github.com/crosswalk-project/crosswalk/pull/2015

BUG=https://crosswalk-project.org/jira/browse/XWALK-1803