Revert "Fix build : Force ld.gold binary usage"
This has caused a regression in Tizen builds from scratch when NaCl
support is enabled (ie. 32-bit Tizen, like IVI and Mobile).

One stage of the NaCl build process involves calling the ld_bfd.py
script, that invokes `gcc -print-prog-name=ld' and later tries finding a
binary with ".bfd" appended to it because certain NaCl binaries must be
built with ld.bfd instead of gold.

binutils's packaging in Tizen is such that it creates symlinks in
/usr/i586-tizen-linux/bin to ar, as, ld, nm, ranlib, strip, but not
ld.bfd or ld.gold. The build-dependency on binutils-gold introduced in
the commit we are reverting makes /usr/bin/ld point to ld.gold instead
of ld.bfd. Finally, when `gcc -print-prog-name=ld' is called it finds
/usr/i586-tizen-linux/bin/ld but not /usr/i586-tizen-linux/bin/ld.bfd,
and since the former points to gold the script ends up linking the NaCl
binaries with gold instead of ld.bfd.

There are different ways to fix this, but for now we have to revert back
to a state where all Tizen profiles build, since this has blocked canary
builds for two days at least.

Related to XWALK-1904.

This reverts commit 55c37c02e29ba6ac21f1bb4031e7b9099ee11a38.